refactored views and api functions to helpers module
