+++ project5/scripts/update_from_api.py
@@ -1,0 +1,6 @@
+from helpers.api_functions import update_data
+
+
+def run():
+    """The function that runs when the script is executed."""
+    update_data()
